+++ include/request_generator.hpp
@@ -12,13 +12,13 @@ namespace palloc {
 
 class Request {
    public:
-    explicit Request(uint64_t dropoffNode, uint64_t requestDuration, uint64_t tillArrival)
+    explicit Request(uint32_t dropoffNode, uint32_t requestDuration, uint32_t tillArrival)
         : _dropoffNode(dropoffNode), _requestDuration(requestDuration), _tillArrival(tillArrival) {}
 
-    uint64_t getDropoffNode() const noexcept;
-    uint64_t getRequestDuration() const noexcept;
-    uint64_t getTimesDropped() const noexcept;
-    uint64_t getArrival() const noexcept;
+    uint32_t getDropoffNode() const noexcept;
+    uint32_t getRequestDuration() const noexcept;
+    uint32_t getTimesDropped() const noexcept;
+    uint32_t getArrival() const noexcept;
 
     void decrementDuration() noexcept;
     void decrementTillArrival() noexcept;
@@ -28,10 +28,10 @@ class Request {
     bool isEarly() const noexcept;
 
    private:
-    uint64_t _dropoffNode;
-    uint64_t _requestDuration;
-    uint64_t _timesDropped = 0;
-    uint64_t _tillArrival;
+    uint32_t _dropoffNode;
+    uint32_t _requestDuration;
+    uint32_t _timesDropped = 0;
+    uint32_t _tillArrival;
 };
 
 using Requests = std::vector<Request>;
@@ -39,10 +39,10 @@ using Requests = std::vector<Request>;
 class RequestGenerator {
    public:
     struct Options {
-        uint64_t dropoffNodes;
-        uint64_t maxTimeTillArrival;
-        uint64_t maxRequestDuration;
-        uint64_t seed;
+        size_t dropoffNodes;
+        uint32_t maxTimeTillArrival;
+        uint32_t maxRequestDuration;
+        uint32_t seed;
         double requestRate;
     };
 
@@ -54,10 +54,10 @@ class RequestGenerator {
           _requestRate(options.requestRate) {
         DoubleVector durationWeights = getDurationBuckets(options.maxRequestDuration);
         _durationDist =
-            std::discrete_distribution<uint64_t>(durationWeights.begin(), durationWeights.end());
+            std::discrete_distribution<uint32_t>(durationWeights.begin(), durationWeights.end());
     }
 
-    Requests generate(uint64_t currentTimeOfDay);
+    Requests generate(uint32_t currentTimeOfDay);
 
    private:
     /**
@@ -66,61 +66,61 @@ class RequestGenerator {
      *
      * @param currentTimeOfDay time of day in minutes from midnight
      */
-    uint64_t getCount(uint64_t currentTimeOfDay);
+    uint32_t getCount(uint32_t currentTimeOfDay);
 
     /**
      * Uniformly sample one of the dropoffs
      * (sample space is warped because they are not uniformly distributed on a map)
      */
-    uint64_t getDropoff();
+    size_t getDropoff();
 
     /**
      * Uniformly sample duration from a random weighted bucket
      */
-    uint64_t getDuration();
+    uint32_t getDuration();
 
     /**
      * Uniformly sample the time till arrival
      */
-    uint64_t getArrival();
+    uint32_t getArrival();
 
     /**
      * Normally poisson is in interval [0, ∞]. When rate > 100 then it becomes a decent
      * approximation of the central limit theorem for gaussian distirbution so we limit it to r
      * ate + 3σ. When rate <= 100 we act like its 100 and limit it to 100 + 3σ
      */
-    static uint64_t getPoissonUpperBound(double rate);
+    static uint32_t getPoissonUpperBound(double rate);
 
     /**
      * Function that returns a multiplier which changes during the day to represent parking requests
      * as a function of time
      */
-    static double getTimeMultiplier(uint64_t currentTimeOfDay);
+    static double getTimeMultiplier(uint32_t currentTimeOfDay);
 
     /**
      * Get viable duration buckets
      */
-    static DoubleVector getDurationBuckets(uint64_t maxDuration);
+    static DoubleVector getDurationBuckets(uint32_t maxDuration);
 
-    std::uniform_int_distribution<uint64_t> _dropoffDist;
-    std::discrete_distribution<uint64_t> _durationDist;
-    std::uniform_int_distribution<uint64_t> _arrivalDist;
+    std::uniform_int_distribution<size_t> _dropoffDist;
+    std::discrete_distribution<uint32_t> _durationDist;
+    std::uniform_int_distribution<uint32_t> _arrivalDist;
     std::minstd_rand _rng;
 
     // Bucket intervals based on COWI
-    static constexpr std::array<std::array<uint64_t, 2>, 7> DURATION_BUCKETS{
+    static constexpr std::array<std::array<uint32_t, 2>, 7> DURATION_BUCKETS{
         {{{0, 60}},                                         // 14%
          {{61, 120}},                                       // 13%
          {{121, 240}},                                      // 10%
          {{241, 480}},                                      // 16%
          {{481, 1440}},                                     // 21%
          {{1441, 2880}},                                    // 9%
-         {{2881, std::numeric_limits<uint64_t>::max()}}}};  // 7%
+         {{2881, std::numeric_limits<uint32_t>::max()}}}};  // 7%
 
     // Weights based on COWI
     static constexpr std::array<double, 7> originalWeights{14.0, 13.0, 10.0, 16.0, 21.0, 9.0, 7.0};
 
-    uint64_t _maxRequestDuration;
+    uint32_t _maxRequestDuration;
     double _requestRate;
 };
 }  // namespace palloc
