+++ include/settings.hpp
@@ -6,23 +6,23 @@
 
 namespace palloc {
 struct SimulatorSettings {
-    uint64_t timesteps;
-    uint64_t startTime;
-    uint64_t maxRequestDuration;
+    uint32_t timesteps;
+    uint32_t startTime;
+    uint32_t maxRequestDuration;
     double requestRate;
-    uint64_t maxTimeTillArrival;
-    uint64_t batchInterval;
-    uint64_t seed;
+    uint32_t maxTimeTillArrival;
+    uint32_t batchInterval;
+    uint32_t seed;
 };
 
 struct OutputSettings {
     std::filesystem::path path;
-    uint64_t numberOfRunsToAggregate;
+    uint32_t numberOfRunsToAggregate;
     bool prettify;
 };
 
 struct GeneralSettings {
-    uint64_t numberOfThreads;
+    uint32_t numberOfThreads;
 };
 }  // namespace palloc
