+++ src/trace.cpp
@@ -8,9 +8,9 @@ size_t Trace::getDroppedRequests() const noexcept { return _droppedRequests; }
 
 size_t Trace::getNumberOfRequests() const noexcept { return _numberOfRequests; }
 
-uint64_t Trace::getAvailableParkingSpots() const noexcept { return _availableParkingSpots; }
+uint32_t Trace::getAvailableParkingSpots() const noexcept { return _availableParkingSpots; }
 
-uint64_t Trace::getTimeStep() const noexcept { return _timestep; }
+uint32_t Trace::getTimeStep() const noexcept { return _timestep; }
 
 double Trace::getCost() const noexcept { return _cost; }
